Move discover_subs_modules function to rele.management.commands.discover
